Refactor to create useFirebaseSnapshot hook and cleanData util function
