+++ src/utils/cleanData.js
@@ -1,0 +1,6 @@
+export default function cleanData(data) {
+  return data
+    .replace(/[.,\/#!$%\^&\*;:{}=\-_`~()@]/g, '')
+    .toLowerCase()
+    .trim();
+}
